feat: update payment

+++ Agents/Agents/Model/Comment.cs
@@ -4,8 +4,8 @@ namespace Agents.Model
 {
     public class Comment:Entity
     {
-        public int ReviewedCompanyId { get; set; }
-        public int UserId { get; set; }
+        public long ReviewedCompanyId { get; set; }
+        public long UserId { get; set; }
 
         public String JobPosition { get; set; }
