+++ Agents/Agents/Model/Payment.cs
@@ -4,8 +4,8 @@ namespace Agents.Model
 {
     public class Payment
     {
-        public int CompanyId { get; set; }
-        public int UserId { get; set; }
+        public long CompanyId { get; set; }
+        public long UserId { get; set; }
         public string JobPosition { get; set; }
         public double Price { get; set; }
